+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- The zOrder is now correctly interpreted in `WebGLSystem` rather than being
+  in order of renderable creation.
 
 ## [v0.8.0] - 2020-05-28
 ### Added